pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = ""
authors = ["Rhuan Emanuel <rhnrod@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-extensions = "^3.2.3"
django-debug-toolbar = "^5.1.0"
psycopg2-binary = "2.9.9"
gitpython = "^3.1.40"  # Adicionado explicitamente

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bookstore.settings"
python_files = ["test*.py", "*_test.py"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
